# contracts/openapi/salas-partidas.yaml
openapi: 3.1.0
...
info:
  title: Nexus Battles VI — Salas y partidas (salas-partidas)
  description: |
    API de creacion, listado e ingreso a salas de batalla, y de consulta del
    estado de una partida en curso.

    El avance del combate turno a turno NO viaja por esta API: viaja por el
    canal WebSocket descrito en `contracts/websocket/salas-partidas.yaml`.
    Esta API cubre el ciclo de vida de la sala hasta el inicio del combate, y
    la consulta puntual de estado para reconexion.

    Errores: todos los errores se devuelven en formato problem details
    (RFC 7807) con `Content-Type: application/problem+json`, identico en los 20
    modulos (regla 4 de plataforma). La correspondencia entre este formato y el
    componente `Aviso` de la interfaz esta en
    `shared/ui-kit/MAPEO-ERRORES.md`.

    Trazas: toda peticion propaga `traceparent` (regla 5 de plataforma).

    Cambios 1.2.0 (HU-SAL-004, aditivos): `heroesIA` en `CrearSalaRequest` y en
    `Sala` — cuantos cupos son de la inteligencia artificial; cuentan en
    `ocupacion`. `CONTRA_IA` lleva siempre uno; `UNO_CONTRA_UNO` ninguno;
    `HASTA_SEIS` de cero a `maximoParticipantes - 1`. `incluirHeroeIA` se
    conserva y vale por uno. En el modo cooperativo (`tamanoEquipo`) los
    participantes de la partida traen `equipo` asignado por orden de entrada.

    Cambios 1.3.0 (HU-DIS-003, degradacion controlada): cuando una dependencia
    de este servicio NO RESPONDE (inventario, motor de combate, libro de
...

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /salas:
    post:
      tags: [Salas]
      summary: Crear una sala de batalla
      description: |
        Crea una sala indicando cuantos jugadores participaran, la recompensa
        en juego y si se incluye un heroe controlado por la inteligencia
        artificial (RF-JUE-001).

        Los creditos de la recompensa quedan comprometidos al crear la sala
        (RF-JUE-014). Si el jugador no tiene saldo suficiente la operacion se
        rechaza con `422` y el `detail` indica cuantos creditos tiene y cuantos
        hacen falta — el requisito exige comunicar el MOTIVO del rechazo, no
        solo que fallo.
      operationId: crearSala
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearSalaRequest'
      responses:
        '201':
          description: Sala creada. El creador queda dentro como anfitrion.
          headers:
            Location:
              description: URI de la sala creada
              schema: { type: string, format: uri }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sala'
        '400':
          $ref: '#/components/responses/ParametrosInvalidos'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/JugadorSancionado'
        '422':
          description: |
            El anfitrion no cumple algun requisito para crear la sala.
            Distinguir por `type`:

            - `creditos-insuficientes`: el saldo no cubre la recompensa;
              `detail` incluye el saldo actual y el requerido.
            - `heroe-no-equipado` / `heroe-ocupado`: RF-JUE-003, SCRUM-1074. El
              anfitrion entra a su propia sala al crearla, asi que pasa la misma
              puerta de heroe que los demas.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              examples:
                creditosInsuficientes:
                  summary: RF-JUE-014 — saldo insuficiente
                  value:
                    type: 'https://nexusbattles.local/errores/creditos-insuficientes'
                    title: 'Creditos insuficientes'
                    status: 422
                    detail: 'Tienes 240 créditos y la apuesta de esta sala es de 400.'
                    instance: '/api/v1/salas'
                heroeNoEquipado:
                  summary: RF-JUE-003 — sin personaje equipado
                  value:
                    type: 'https://nexusbattles.local/errores/heroe-no-equipado'
                    title: 'No tienes un héroe equipado'
                    status: 422
                    detail: 'Equipa un héroe en tu inventario antes de entrar a una batalla.'
                    instance: '/api/v1/salas'
        '503':
          $ref: '#/components/responses/DependenciaNoDisponible'

    # -------------------------------------------------------------------------
    # RF-JUE-002 · HU-SAL-002 — Listado de salas creadas por otros jugadores
    # -------------------------------------------------------------------------
    get:
      tags: [Salas]
      summary: Listar salas disponibles
      description: |
        Devuelve la lista de salas creadas por otros jugadores para que el
        jugador seleccione una e ingrese (RF-JUE-002).

        La paginacion es de 16 elementos por pagina, valor fijado en el sistema
        de diseno y confirmado por el componente `Paginacion`, cuya descripcion
        cita RNF-USA-001: «dieciseis elementos por pagina en vitrinas y
        listados».

        LAS SALAS PRIVADAS SI APARECEN. La Pantalla 2 del sistema de diseno
        (nodo 19:29) pinta ocho tarjetas y dos llevan la insignia `Privada`.
        Aparecer no es poder entrar: el ingreso a una privada sin codigo se
        rechaza con `403` en `POST /salas/{idSala}/participantes`. Si no
        aparecieran, ese 403 seria inalcanzable desde el listado.
      operationId: listarSalas
      parameters:
        - name: pagina
          in: query
          description: Numero de pagina, base 0
          schema: { type: integer, minimum: 0, default: 0 }
        - name: tamano
          in: query
          description: Elementos por pagina. El valor de diseno es 16.
          schema: { type: integer, minimum: 1, maximum: 50, default: 16 }
        - name: modalidad
          in: query
          description: Filtro por modalidad de partida (RF-JUE-004)
          schema:
            $ref: '#/components/schemas/Modalidad'
        - name: estado
          in: query
          description: |
            Filtro por estado de la sala. Desde 1.10.0 se puede repetir o separar
            con comas: `ABIERTA,PRIVADA` son las salas a las que se puede entrar,
            lo que «Jugar online» pide por omision. Un estado que no aparece en
            el listado (`EN_JUEGO`, `CANCELADA`, `FINALIZADA`) no devuelve nada
            aunque se pida. Sin el parametro no se filtra. Las salas salen
            siempre de la mas reciente a la mas antigua.
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/EstadoSala'
        # RETIRADO: buscaba por nombre de sala, y las salas no tienen nombre.
        # Ningun requisito pide buscar en el listado; RF-JUE-002 solo habla de
        # seleccionar una sala de la lista.
        #
        # RESUELTO EN HU-SAL-002. La Pantalla 2 conservaba un campo «Buscar una
        # sala por nombre...» que quedo huerfano al retirar `Nombre` de la
        # tarjeta en HU-SAL-001: ninguna de sus ocho tarjetas mostraba un nombre,
        # porque la propiedad `Meta` de `Tarjeta de sala` solo lleva
        # participantes, recompensa y heroe de la IA. El campo se elimino de la
        # maqueta, no solo del codigo, para que el diseno no prometa una funcion
        # que no existe. Quedan los dos filtros que si tienen respaldo:
        # modalidad y estado.
      responses:
        '200':
          description: Pagina de salas disponibles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginaDeSalas'
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /salas/{idSala}:
    parameters:
      - $ref: '#/components/parameters/IdSala'
    get:
      tags: [Salas]
      summary: Consultar una sala
      operationId: obtenerSala
      responses:
        '200':
          description: Detalle de la sala
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sala'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '404':
          $ref: '#/components/responses/SalaNoEncontrada'

    delete:
      tags: [Salas]
      summary: Cancelar una sala
      description: |
        Solo el anfitrion puede cancelar. Los creditos comprometidos por los
        participantes se devuelven. Emite el evento `sala.cancelada` por el
        canal en tiempo real para que quienes esten dentro salgan de la vista
        de espera.
      operationId: cancelarSala
      responses:
        '204':
          description: Sala cancelada y creditos devueltos
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          description: Solo el anfitrion puede cancelar la sala
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '404':
          $ref: '#/components/responses/SalaNoEncontrada'
        '409':
          description: La partida ya comenzo; no se puede cancelar
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }

  # ---------------------------------------------------------------------------
  # RF-JUE-002 + RF-JUE-003 · HU-SAL-002 / HU-SAL-003 — Ingreso con validacion
  # ---------------------------------------------------------------------------
  /salas/{idSala}/participantes:
    parameters:
      - $ref: '#/components/parameters/IdSala'
    post:
      tags: [Salas]
      summary: Ingresar a una sala
      description: |
        Ingresa al jugador autenticado en la sala seleccionada (RF-JUE-002).

        1.9.0 — orden de las comprobaciones: sancion → la sala existe → la
        sala admite al jugador con su codigo (`comprobarIngreso`) → heroe →
        creditos. Un codigo equivocado o una sala llena se dicen sin preguntar
        al inventario ni reservar nada.

        Antes de admitirlo, el sistema exige que tenga su personaje equipado
        (RF-JUE-003) y que ese heroe no este comprometido en otra partida. Los
        dos rechazos posibles se distinguen por `type`, porque la interfaz
        muestra un dialogo distinto para cada uno:

          - `heroe-no-equipado`  → invita a ir al inventario
          - `heroe-ocupado`      → nombra la sala donde esta y ofrece cambiarlo

        Al ingresar se comprometen los creditos de la apuesta (RF-JUE-014).
      operationId: ingresarASala
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngresoRequest'
      responses:
        '200':
          description: Ingreso aceptado. Devuelve la sala con el participante dentro.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sala'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          description: |
            No puede entrar. Distinguir por `type`:

            - `sala-privada`: la sala es privada y el codigo falta o no
              corresponde. (1.9.0: el ejemplo decia `codigo-invitacion-invalido`,
              un tipo que el servicio nunca emitio; el real es este, el que
              la interfaz siempre ha leido.)
            - `jugador-sancionado` (HU-USR-005/006, R10.2): tiene una sancion
              activa. Se comprueba ANTES que el heroe y antes de reservar
              creditos, asi que un rechazo por esta via no deja nada a medias.
              Ver la respuesta `JugadorSancionado`.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
              examples:
                codigoInvalido:
                  summary: sala privada sin codigo valido
                  value:
                    type: 'https://nexusbattles.local/errores/sala-privada'
                    title: 'Codigo de invitacion invalido'
                    status: 403
                    detail: 'Esta sala es privada y el codigo no corresponde.'
                jugadorSancionado:
                  summary: HU-USR-005/006 — sancion activa
                  value:
                    type: 'https://nexusbattles.local/errores/jugador-sancionado'
                    title: 'Tienes una sanción activa'
                    status: 403
                    detail: 'Mientras la sanción siga vigente no puedes crear salas ni entrar a batallas. Puedes consultarla, y apelarla si crees que es injusta, en «Mis sanciones».'
        '404':
          $ref: '#/components/responses/SalaNoEncontrada'
        '409':
          description: La sala esta llena o la partida ya comenzo
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '422':
          description: |
            El jugador no cumple algun requisito para entrar. Distinguir por
            `type`:

            - `heroe-no-equipado` / `heroe-ocupado`: RF-JUE-003 (ver
              `HeroeNoDisponible`).
            - `creditos-insuficientes`: RF-JUE-014, HU-JUE-014 CA-02. La sala
              tiene recompensa y el saldo disponible no la cubre; `detail`
              incluye el saldo actual y el requerido. La sala no cambia y
              nada queda reservado.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              examples:
                creditosInsuficientes:
                  value:
                    type: 'https://nexusbattles.local/errores/creditos-insuficientes'
                    title: 'Creditos insuficientes'
                    status: 422
                    detail: 'Tienes 80 créditos y la apuesta de esta sala es de 150.'
                heroeNoEquipado:
                  value:
                    type: 'https://nexusbattles.local/errores/heroe-no-equipado'
                    title: 'No tienes un héroe equipado'
                    status: 422
                    detail: 'Equipa un héroe en tu inventario antes de entrar a una batalla.'
        '503':
          $ref: '#/components/responses/DependenciaNoDisponible'

    delete:
      tags: [Salas]
      summary: Abandonar la sala
      description: |
        Saca al jugador autenticado de la sala antes de que comience la
        partida y le devuelve los creditos comprometidos. Una vez iniciado el
        combate esta ruta ya no sirve (`409`): salir de un combate es
        rendirse, `POST /partidas/{idPartida}/rendicion` (1.10.0), que cuenta
        como derrota y liquida la apuesta con las reglas de la partida.
      operationId: abandonarSala
      responses:
        '204':
          description: Jugador retirado de la sala
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '404':
          $ref: '#/components/responses/SalaNoEncontrada'

  # ---------------------------------------------------------------------------
  # 1.10.0 · Revision del modo jugador (6-oct, punto 13) — Invitar por apodo
  # ---------------------------------------------------------------------------
...
  /salas/{idSala}/verificacion-heroe:
    parameters:
      - $ref: '#/components/parameters/IdSala'
    get:
      tags: [Salas]
      summary: Verificar el heroe antes de intentar ingresar
      description: |
        Comprueba si el jugador puede entrar, SIN ingresarlo ni comprometer
        creditos. La interfaz la usa para mostrar el dialogo de validacion de
        heroe antes de pulsar Entrar (RF-JUE-003), de modo que el jugador no
        descubra el problema despues de la accion.
      operationId: verificarHeroe
      responses:
        '200':
          description: Resultado de la verificacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificacionHeroe'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '404':
          $ref: '#/components/responses/SalaNoEncontrada'
        '503':
          # La verificacion pregunta al inventario: si no responde, no hay
          # veredicto (ni inventado). El dialogo pinta `Seccion degradada`
          # con reintentar en vez de una de sus cuatro variantes.
          $ref: '#/components/responses/DependenciaNoDisponible'

  # ---------------------------------------------------------------------------
  # HU-SAL-004 · RF-JUE-017 — Arranque del combate
  # ---------------------------------------------------------------------------
...
components:
...
  responses:
...
    NoAutenticado:
      description: Falta el token o esta caducado
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    # HU-DIS-003 — la misma forma en toda operacion que necesita a otro servicio.
...
    ParametrosInvalidos:
      description: |
        Parametros fuera de rango. `errores` detalla campo por campo para que
        la interfaz marque el campo concreto, no solo el formulario.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
  schemas:
...
    Modalidad:
      type: string
      description: |
        Modalidades admitidas (RF-JUE-004): uno contra uno, uno contra la
        inteligencia artificial, y partidas de hasta seis jugadores.

        Limites por modalidad (los aplica el servicio; fuera de ellos, `400`
        con el campo y el limite superado):

        | Modalidad      | maximoParticipantes | heroesIA                 | tamanoEquipo |
        |----------------|---------------------|--------------------------|--------------|
        | UNO_CONTRA_UNO | exactamente 2       | 0                        | no admite    |
        | CONTRA_IA      | exactamente 2       | exactamente 1 (implicito)| no admite    |
        | HASTA_SEIS     | 2 a 6               | 0 a maximoParticipantes-1| 1 a 3        |
      enum: [UNO_CONTRA_UNO, CONTRA_IA, HASTA_SEIS]
    EstadoSala:
      type: string
      description: |
        Estado de la sala. Se corresponde uno a uno con las variantes del
        componente `Insignia` en el sistema de diseno.
      enum: [ABIERTA, LLENA, EN_JUEGO, PRIVADA, CANCELADA, FINALIZADA]
...
  securitySchemes:
    keycloak:
      type: openIdConnect
      openIdConnectUrl: /realms/nexus-battles/.well-known/openid-configuration
      description: Identidad emitida por Keycloak 26.
security:
  - keycloak: []