# services/plataforma/salas-partidas/src/main/java/com/nexusbattles/plataforma/salaspartidas/api/SalasController.java
package com.nexusbattles.plataforma.salaspartidas.api;

import com.nexusbattles.plataforma.salaspartidas.aplicacion.AbandonarSala;
import com.nexusbattles.plataforma.salaspartidas.aplicacion.CancelarSala;
import com.nexusbattles.plataforma.salaspartidas.aplicacion.ComprobarIngreso;
import com.nexusbattles.plataforma.salaspartidas.aplicacion.CrearSala;
import com.nexusbattles.plataforma.salaspartidas.aplicacion.InformarEncuentroDeTorneo;
import com.nexusbattles.plataforma.salaspartidas.aplicacion.IngresarASala;
import com.nexusbattles.plataforma.salaspartidas.aplicacion.IniciarPartida;
import com.nexusbattles.plataforma.salaspartidas.aplicacion.InvitarASala;
import com.nexusbattles.plataforma.salaspartidas.aplicacion.JugadorAutenticado;
import com.nexusbattles.plataforma.salaspartidas.aplicacion.ListarSalas;
import com.nexusbattles.plataforma.salaspartidas.aplicacion.ObtenerSala;
import com.nexusbattles.plataforma.salaspartidas.aplicacion.VerificarHeroe;
import com.nexusbattles.plataforma.salaspartidas.dominio.EstadoSala;
import com.nexusbattles.plataforma.salaspartidas.dominio.Modalidad;
import com.nexusbattles.plataforma.salaspartidas.dominio.Sala;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.security.oauth2.jwt.Jwt;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.util.UriComponentsBuilder;

import java.util.UUID;

/**
 * API de salas de batalla — HU-SAL-001 y HU-SAL-002.
 *
 * <p>Regla 2 de plataforma: todo bajo {@code /api/v1}. Un cambio incompatible
 * abre version nueva, no modifica esta.
 *
 * <p>Ni el anfitrion ni el jugador que entra viajan en el cuerpo: salen del
 * token. Si vinieran del cliente, cualquiera podria crear salas o entrar a
 * nombre de otro cambiando un campo del JSON.
 */
@RestController
@RequestMapping("/api/v1/salas")
public class SalasController {

    private final CrearSala crearSala;
    private final ListarSalas listarSalas;
    private final IngresarASala ingresarASala;
    private final ObtenerSala obtenerSala;
    private final AbandonarSala abandonarSala;
    private final CancelarSala cancelarSala;
    private final VerificarHeroe verificarHeroe;
    private final IniciarPartida iniciarPartida;
    private final InformarEncuentroDeTorneo encuentroDeTorneo;
    private final ComprobarIngreso comprobarIngreso;
    private final InvitarASala invitarASala;

    SalasController(CrearSala crearSala, ListarSalas listarSalas, IngresarASala ingresarASala,
                    ObtenerSala obtenerSala, AbandonarSala abandonarSala,
                    CancelarSala cancelarSala, VerificarHeroe verificarHeroe,
                    IniciarPartida iniciarPartida, InformarEncuentroDeTorneo encuentroDeTorneo,
                    ComprobarIngreso comprobarIngreso, InvitarASala invitarASala) {
        this.comprobarIngreso = comprobarIngreso;
        this.invitarASala = invitarASala;
        this.iniciarPartida = iniciarPartida;
        this.encuentroDeTorneo = encuentroDeTorneo;
        this.crearSala = crearSala;
        this.listarSalas = listarSalas;
        this.ingresarASala = ingresarASala;
        this.obtenerSala = obtenerSala;
        this.abandonarSala = abandonarSala;
        this.cancelarSala = cancelarSala;
        this.verificarHeroe = verificarHeroe;
    }

    /**
     * Crea una sala de batalla (RF-JUE-001).
     *
     * @return 201 con la sala creada y su cabecera {@code Location}
     */
    @PostMapping
    public ResponseEntity<SalaResponse> crear(@RequestBody CrearSalaRequest peticion,
                                              @AuthenticationPrincipal Jwt token) {

        JugadorAutenticado anfitrion = jugadorDe(token);
        if (peticion.torneo() != null) {
            // Se valida ANTES de crear la sala: un encuentro fuera de rango no
            // deja una sala huerfana. El numero nulo cae en el mismo 400.
            validarEncuentro(peticion.torneo());
        }
        Sala sala = crearSala.ejecutar(peticion.aParametros(), anfitrion);
        if (peticion.torneo() != null) {
            encuentroDeTorneo.vincular(sala.id(), peticion.torneo().torneoId(),
                    peticion.torneo().numeroEncuentro(), anfitrion.id());
        }

        return ResponseEntity
                .created(UriComponentsBuilder.fromPath("/api/v1/salas/{id}")
                        .buildAndExpand(sala.id())
                        .toUri())
                // Quien crea la sala es su anfitrion: esta es la unica respuesta
                // en la que el codigo de invitacion se entrega sin pedirlo, y es
                // la que hace utilizable una sala privada.
                .body(SalaResponse.paraElAnfitrion(sala));
    }

    /**
     * Lista las salas para que el jugador elija una (RF-JUE-002).
     *
     * <p>Los cuatro parametros se pasan tal cual, incluso nulos: los valores por
     * defecto los decide el caso de uso, que es donde vive esa regla. Si el
     * controlador rellenara aqui el 16, habria dos sitios donde cambiarlo.
     *
     * <p>1.10.0 — {@code estado} se puede repetir ({@code ?estado=ABIERTA&estado=PRIVADA})
     * o separar con comas: las salas a las que se puede entrar son las dos.
     */
    @GetMapping
    public PaginaDeSalasResponse listar(@RequestParam(required = false) Integer pagina,
                                        @RequestParam(required = false) Integer tamano,
                                        @RequestParam(required = false) Modalidad modalidad,
                                        @RequestParam(required = false) java.util.List<EstadoSala> estado) {

        java.util.Set<EstadoSala> estados = estado == null
                ? java.util.Set.of()
                : estado.stream().filter(java.util.Objects::nonNull)
                        .collect(java.util.stream.Collectors.toCollection(
                                () -> java.util.EnumSet.noneOf(EstadoSala.class)));
        return PaginaDeSalasResponse.desde(listarSalas.ejecutarEnEstados(pagina, tamano, modalidad, estados));
    }

    /**
     * Ingresa al jugador autenticado en la sala elegida (RF-JUE-002).
     *
     * <p>Los tres rechazos posibles los traduce {@code ManejadorDeErrores} a los
     * codigos que fija el contrato, y la interfaz los distingue por el campo
     * {@code type}: 404 la sala no existe, 403 es privada y falta la
     * invitacion, 409 esta llena o la partida ya empezo.
     */
    @PostMapping("/{idSala}/participantes")
    public SalaResponse ingresar(@PathVariable UUID idSala,
                                 @RequestBody(required = false) IngresoRequest peticion,
                                 @AuthenticationPrincipal Jwt token) {

        JugadorAutenticado jugador = jugadorDe(token);
        Sala sala = ingresarASala.ejecutar(idSala, jugador, IngresoRequest.codigoDe(peticion));
        return SalaResponse.segunQuienPregunta(sala, jugador.id());
    }

    /**
     * ¿Me dejaria entrar? — operacion {@code comprobarIngreso} (1.9.0, RFINAL-04).
     *
     * <p>Sin efectos: ni ingresa, ni reserva, ni pregunta al inventario. El
     * listado la usa antes de mandar a verificar el heroe, para que un codigo
     * de invitacion equivocado se diga ahi (403) y no despues de elegir heroe.
     * Los rechazos son los del ingreso y los traduce el mismo
     * {@code ManejadorDeErrores}. El codigo viaja en el cuerpo, no en la URL:
     * la URL acaba en las bitacoras del borde.
     *
     * @return 200 con la sala, vista como la ve quien pregunta
     */
    @PostMapping("/{idSala}/comprobacion-de-ingreso")
    public SalaResponse comprobarIngreso(@PathVariable UUID idSala,
                                         @RequestBody(required = false) IngresoRequest peticion,
                                         @AuthenticationPrincipal Jwt token) {

        JugadorAutenticado jugador = jugadorDe(token);
        Sala sala = comprobarIngreso.ejecutar(idSala, jugador, IngresoRequest.codigoDe(peticion));
        return SalaResponse.segunQuienPregunta(sala, jugador.id());
    }

    /**
     * Devuelve una sala concreta (operacion {@code obtenerSala}).
     *
     * <p>Es el destino de la cabecera {@code Location} de la creacion y lo que
     * necesita la vista de sala cuando se llega por enlace directo, sin pasar
     * por el listado.
     *
     * <p>Al anfitrion le llega ademas el codigo de invitacion, para que pueda
     * volver a consultarlo sin tener que guardar la respuesta de la creacion.
     */
    @GetMapping("/{idSala}")
    public SalaResponse obtener(@PathVariable UUID idSala,
                                @AuthenticationPrincipal Jwt token) {

        UUID idJugador = idDe(token);
        Sala sala = obtenerSala.ejecutar(idSala);
        // R18 — con su partida si ya arranco: sin ella, recargar a mitad de
        // combate devolvia a la sala de espera y no habia forma de volver.
        UUID idPartida = obtenerSala.partidaDe(sala).orElse(null);
        return SalaResponse.segunQuienPregunta(sala, idJugador, idPartida);
    }

    /**
     * Cancela la sala (operacion {@code cancelarSala}).
     *
     * <p>Solo el anfitrion: 403 si lo pide otro, 409 si la partida ya arranco.
     * Devuelve 204 porque despues de cancelarla no queda nada util que
     * representar — el estado final ya viaja por el canal, a todos los que
     * estaban dentro, en el aviso {@code sala.cancelada}.
     */
    @DeleteMapping("/{idSala}")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void cancelar(@PathVariable UUID idSala, @AuthenticationPrincipal Jwt token) {
        cancelarSala.ejecutar(idSala, idDe(token));
    }

    /**
     * Saca de la sala al jugador autenticado (operacion {@code abandonarSala}).
     *
     * <p>El anfitrion no usa este camino: recibe 409 y se le remite a cancelar.
     * Ver {@code Sala#abandonar}.
     */
    @DeleteMapping("/{idSala}/participantes")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void abandonar(@PathVariable UUID idSala, @AuthenticationPrincipal Jwt token) {
        abandonarSala.ejecutar(idSala, idDe(token));
    }

    /**
     * Invita a un jugador a la sala — operacion {@code invitarASala} (1.10.0).
     *
     * <p>Solo el anfitrion. El invitado se elige en la vista buscandolo por su
     * apodo ({@code GET /perfiles/publicos} de ms-identidad) y viaja aqui por su
     * identificador, que el jugador nunca ve. Le llega un aviso a su bandeja con
     * la sala y, si es privada, el codigo. Los rechazos son
     * {@code invitacion-no-permitida} con el motivo en {@code detail}.
     *
     * @return 200 con a quien se invito y si el aviso salio ahora ({@code enviada})
     *         o ya lo tenia de antes
     */
    @PostMapping("/{idSala}/invitaciones")
    public InvitacionResponse invitar(@PathVariable UUID idSala,
                                      @RequestBody(required = false) InvitacionRequest peticion,
                                      @AuthenticationPrincipal Jwt token) {

        if (peticion == null || peticion.idJugador() == null) {
            throw new com.nexusbattles.plataforma.salaspartidas.dominio.ParametrosInvalidos(
                    "idJugador", "Falta el jugador al que invitas");
        }
        var resultado = invitarASala.ejecutar(idSala, jugadorDe(token), peticion.idJugador());
        return new InvitacionResponse(resultado.idJugador(), resultado.apodo(), resultado.enviada());
    }

    /** Cuerpo de {@code invitarASala}: a quien. */
    public record InvitacionRequest(UUID idJugador) {
    }

    /** Respuesta de {@code invitarASala}. */
    public record InvitacionResponse(UUID idJugador, String apodo, boolean enviada) {
    }

    /**
     * Arranca el combate de la sala (HU-SAL-004, RF-JUE-017).
     *
     * <p>Solo el anfitrion: 403 si lo pide otro, 409 si la sala ya no admite
     * empezar —cancelada, terminada, o sin rival—. Devuelve 201 con el estado
     * inicial de la partida y su {@code Location}, para que el anfitrion no
     * tenga que volver a preguntarlo.
     *
     * <p>A los demas participantes les llega por el canal de la sala el aviso
     * {@code sala.partida.iniciada}: es lo que los mueve de la sala de espera al
     * combate sin recargar.
     *
     * <p>Pulsar dos veces no es un error: la segunda llamada devuelve la misma
     * partida.
     */
    @PostMapping("/{idSala}/partida")
    public ResponseEntity<PartidaResponse> iniciar(@PathVariable UUID idSala,
                                                   @AuthenticationPrincipal Jwt token) {

        var partida = iniciarPartida.ejecutar(idSala, jugadorDe(token));

        return ResponseEntity
                .created(UriComponentsBuilder.fromPath("/api/v1/partidas/{id}")
                        .buildAndExpand(partida.id())
                        .toUri())
                .body(PartidaResponse.desde(partida));
    }

    /**
     * Verifica el heroe antes de que el jugador pulse Entrar (HU-SAL-003,
     * RF-JUE-003).
     *
     * <p>No tiene efectos: ni ingresa, ni reserva, ni bloquea nada. Solo
     * responde, para que el dialogo de validacion diga el motivo <b>antes</b> de
     * la accion y no despues. 404 si la sala no existe; 503 si el inventario no
     * contesta, porque un veredicto inventado mandaria al jugador a chocar
     * contra el rechazo que esta ruta existe para evitar.
     */
    @GetMapping("/{idSala}/verificacion-heroe")
    public VerificacionHeroeResponse verificarHeroe(@PathVariable UUID idSala,
                                                    @AuthenticationPrincipal Jwt token) {

        return VerificacionHeroeResponse.desde(
                verificarHeroe.ejecutar(idSala, jugadorDe(token)));
    }

    /**
     * El identificador estable del jugador, nunca un dato del cuerpo.
     *
     * <p>Sale de {@code uid} cuando el token lo trae, y solo si no, del sujeto.
     * El {@code sub} que emite {@code ms-identidad} es el <b>apodo</b> —mutable
     * y no un UUID—, asi que leer el sujeto a secas hacia reventar
     * {@code UUID.fromString} y devolver 500 en cuanto alguien creaba una sala
     * con una sesion de verdad: el listado funcionaba y la creacion no.
     *
     * <p>Es el mismo criterio que aplica {@code ConversorRolesJwt} al elegir el
     * nombre del principal, y sigue sirviendo para un token de Keycloak, donde
     * el sujeto si es el identificador y {@code uid} no existe.
     */
    private static UUID idDe(Jwt token) {
        String uid = token.getClaimAsString("uid");
        return UUID.fromString(uid != null && !uid.isBlank() ? uid : token.getSubject());
    }

    /**
     * Las dos caras de la identidad, tal como las trae el token.
     *
     * <p>El identificador estable manda dentro de este servicio. El apodo solo
     * hace falta para preguntarle al inventario, que hoy reconoce al jugador por
     * {@code X-User-Name} (ver {@code JugadorAutenticado}).
     *
     * <p>Se busca en {@code preferred_username} —el nombre estandar en OIDC, y
     * el que emite Keycloak— y si no esta, en {@code apodo}. Como ultimo
     * recurso queda el sujeto: en los tokens de {@code ms-identidad} anteriores
     * a ADR-002 el sujeto <i>era</i> el apodo.
     */
    private static void validarEncuentro(CrearSalaRequest.EncuentroDeTorneoRequest encuentro) {
        if (encuentro.torneoId() == null) {
            throw new com.nexusbattles.plataforma.salaspartidas.dominio.ParametrosInvalidos(
                    "torneo.torneoId", "Falta el torneo del encuentro");
        }
        Integer numero = encuentro.numeroEncuentro();
        if (numero == null || numero < 1 || numero > 14) {
            throw new com.nexusbattles.plataforma.salaspartidas.dominio.ParametrosInvalidos(
                    "torneo.numeroEncuentro", "El encuentro de un torneo va del 1 al 14");
        }
    }

    private static JugadorAutenticado jugadorDe(Jwt token) {
        return new JugadorAutenticado(
                com.nexusbattles.comun.seguridad.IdentidadDelToken.idDe(token),
                com.nexusbattles.comun.seguridad.IdentidadDelToken.apodoDe(token));
    }

    private static String primerTextoNoVacio(String... candidatos) {
        for (String candidato : candidatos) {
            if (candidato != null && !candidato.isBlank()) {
                return candidato;
            }
        }
        throw new IllegalStateException("El token no trae ningun nombre con el que identificar al jugador.");
    }
}